flowengine-app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
anthropic==0.40.0
sqlalchemy==2.0.23
pytest==7.4.3
httpx==0.27.0
google-generativeai==0.3.2
